suryantara terbaru
tolong cepat dan benar ya :smile:

+++ resources/views/admin/index.blade.php
@@ -12,7 +12,14 @@
 				</div>
 			</header>
 			<section class="card">
+				   
 				<div class="card-block">
+					 @if (session('status'))
+                    <div class="alert alert-success alert-dismissable">
+                                <button type="button" class="close" data-dismiss="alert" aria-hidden="true">&times;</button>
+                                {{ session('status') }}
+                    </div>
+                    @endif
 					<a href="{{url('admin/create')}}" class="btn btn-primary"><i class="fa fa-pencil"></i> Tambah Data</a>
                     <br><br>
 					<table id="example" class="display table table-striped table-bordered" cellspacing="0" width="100%">
@@ -22,7 +29,6 @@
 							<th>Kode Admin</th>
 							<th>Nama</th>
 							<th>Username</th>
-							<th>Password</th>
 							<th>Aksi</th>
 						</tr>
 						</thead>
@@ -32,7 +38,6 @@
 							<th>kode Admin</th>
 							<th>Nama</th>
 							<th>Username</th>
-							<th>Password</th>
 							<th>Aksi</th>
 						</tr>
 						</tfoot>
@@ -45,11 +50,10 @@
                             <td>{{$row->kode}}</td>
                             <td>{{$row->nama}}</td>
                             <td>{{$row->username}}</td>
-                            <td>{{$row->password}}</td>
                             <td><a href="admin/{{$row->id}}" class="btn btn-warning btn-sm">
-                                        <i class="fa fa-key"></i> Ganti Password</a>
-                                <a href="admin/{{$row->id}}/delete" class="btn btn-danger btn-sm">
-                                        <i class="fa fa-remove"></i>hapus</a>
+                                        <i class="fa fa-key"></i> Edit Data</a>
+                                <a href="admin/{{$row->id}}/delete" class="btn btn-danger btn-sm" onclick="return confirm('Hapus Data ?')">
+                                        <i class="fa fa-remove" onclick="return confirm('Hapus Data ?')"></i>hapus</a>
                             </td>
 						</tr>
 						@endforeach
