+++ resources/views/trflaut/index.blade.php
@@ -13,6 +13,12 @@
 			</header>
 			<section class="card">
 				<div class="card-block">
+					@if (session('status'))
+                    <div class="alert alert-success alert-dismissable">
+                                <button type="button" class="close" data-dismiss="alert" aria-hidden="true">&times;</button>
+                                {{ session('status') }}
+                    </div>
+                    @endif
 					<a href="{{url('trflaut/create')}}" class="btn btn-primary"><i class="fa fa-pencil"></i> Tambah Data</a>
                     <br><br>
 					<table id="example" class="display table table-striped table-bordered" cellspacing="0" width="100%">
@@ -45,8 +51,8 @@
                             <td>{{$row->tarif}}</td>
                             <td><a href="trflaut/{{$row->id}}" class="btn btn-warning btn-sm">
                                         <i class="fa fa-pencil"></i> Edit</a>
-                                <a href="trflaut/{{$row->id}}/delete" class="btn btn-danger btn-sm">
-                                        <i class="fa fa-remove"></i>Hapus</a>
+                                <a href="trflaut/{{$row->id}}/delete" class="btn btn-danger btn-sm" onclick="return confirm('Hapus Data ?')">
+                                        <i class="fa fa-remove" onclick="return confirm('Hapus Data ?')"></i>Hapus</a>
                             </td>
 						</tr>
 						@endforeach
